requirements.txt
typer[all]
rich
httpx
Pillow
pydantic
PyQt6

